requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pycryptodome==3.19.0
plotly==5.18.0
kaleido==0.2.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
